+++ examples/with-markdown-plugins/astro.config.mjs
@@ -7,10 +7,10 @@ export default defineConfig({
 	markdown: {
 		remarkPlugins: ['remark-code-titles'],
 		rehypePlugins: [
+			'rehype-slug',
 			['rehype-autolink-headings', { behavior: 'prepend' }],
 			['rehype-toc', { headings: ['h2', 'h3'] }],
 			[addClasses, { 'h1,h2,h3': 'title' }],
-			'rehype-slug',
 		],
 	},
 });