+++ packages/astro/test/0-css.test.js
@@ -5,7 +5,7 @@
  */
 
 import { expect } from 'chai';
-import cheerio from 'cheerio';
+import * as cheerio from 'cheerio';
 import { loadFixture } from './test-utils.js';
 
 let fixture;