+++ packages/astro/test/astro-directives.test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import cheerio from 'cheerio';
+import * as cheerio from 'cheerio';
 import { loadFixture } from './test-utils.js';
 
 describe('Directives', async () => {
@@ -16,6 +16,24 @@ describe('Directives', async () => {
 
 		expect($('script#inline')).to.have.lengthOf(1);
 		expect($('script#inline').toString()).to.include('let foo = "bar"');
+	});
+
+	it('Passes define:vars to style elements', async () => {
+		const html = await fixture.readFile('/define-vars/index.html');
+		const $ = cheerio.load(html);
+
+		expect($('style')).to.have.lengthOf(1);
+		expect($('style').toString()).to.include('--bg: white;');
+		expect($('style').toString()).to.include('--fg: black;');
+
+		const scopedClass = $('html')
+			.attr('class')
+			.split(' ')
+			.find((name) => /^astro-[A-Za-z0-9-]+/.test(name));
+
+		expect($('style').toString().replace(/\s+/g, '')).to.equal(
+			`<style>.${scopedClass}{--bg:white;--fg:black;}body{background:var(--bg);color:var(--fg)}</style>`
+		);
 	});
 
 	it('set:html', async () => {