+++ packages/astro/test/static-build-frameworks.test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import cheerio from 'cheerio';
+import * as cheerio from 'cheerio';
 import { loadFixture, isWindows } from './test-utils.js';
 
 describe('Static build - frameworks', () => {