+++ packages/create-astro/src/index.ts
@@ -1,15 +1,13 @@
 import fs from 'fs';
 import path from 'path';
-import { bold, cyan, gray, green, red, yellow } from 'kleur/colors';
-import fetch from 'node-fetch';
+import { bgCyan, black, bold, cyan, gray, green, red, yellow } from 'kleur/colors';
 import prompts from 'prompts';
 import degit from 'degit';
 import yargs from 'yargs-parser';
 import ora from 'ora';
-import { FRAMEWORKS, COUNTER_COMPONENTS, Integration } from './frameworks.js';
 import { TEMPLATES } from './templates.js';
-import { createConfig } from './config.js';
 import { logger, defaultLogLevel } from './logger.js';
+import { execa, execaCommand } from 'execa';
 
 // NOTE: In the v7.x version of npm, the default behavior of `npm init` was changed
 // to no longer require `--` to pass args and instead pass `--` directly to us. This
@@ -36,17 +34,13 @@ const { version } = JSON.parse(
 	fs.readFileSync(new URL('../package.json', import.meta.url), 'utf-8')
 );
 
-const FILES_TO_REMOVE = ['.stackblitzrc', 'sandbox.config.json']; // some files are only needed for online editors when using astro.new. Remove for create-astro installs.
-const POSTPROCESS_FILES = ['package.json', 'astro.config.mjs', 'CHANGELOG.md']; // some files need processing after copying.
+const FILES_TO_REMOVE = ['.stackblitzrc', 'sandbox.config.json', 'CHANGELOG.md']; // some files are only needed for online editors when using astro.new. Remove for create-astro installs.
 
 export async function main() {
+	const pkgManager = pkgManagerFromUserAgent(process.env.npm_config_user_agent);
+
 	logger.debug('Verbose logging turned on');
 	console.log(`\n${bold('Welcome to Astro!')} ${gray(`(create-astro v${version})`)}`);
-	console.log(
-		`If you encounter a problem, visit ${cyan(
-			'https://github.com/withastro/astro/issues'
-		)} to search or file a new issue.\n`
-	);
 
 	let spinner = ora({ color: 'green', text: 'Prepare for liftoff.' });
 
@@ -63,8 +57,7 @@ export async function main() {
 	}
 
 	if (!cwd || !isEmpty(cwd)) {
-		const notEmptyMsg = (dirPath: string) =>
-			`"${bold(dirPath)}" is not empty. Please clear contents or choose a different path.`;
+		const notEmptyMsg = (dirPath: string) => `"${bold(dirPath)}" is not empty!`;
 
 		if (!isEmpty(cwd)) {
 			let rejectProjectDir = ora({ color: 'red', text: notEmptyMsg(cwd) });
@@ -104,9 +97,7 @@ export async function main() {
 
 	const hash = args.commit ? `#${args.commit}` : '';
 
-	const templateTarget = options.template.includes('/')
-		? options.template
-		: `withastro/astro/examples/${options.template}#latest`;
+	const templateTarget = `withastro/astro/examples/${options.template}#latest`;
 
 	const emitter = degit(`${templateTarget}${hash}`, {
 		cache: false,
@@ -120,173 +111,168 @@ export async function main() {
 		verbose: defaultLogLevel === 'debug' ? true : false,
 	});
 
-	const selectedTemplate = TEMPLATES.find((template) => template.value === options.template);
-	let integrations: Integration[] = [];
-
-	if (selectedTemplate?.integrations === true) {
-		const result = await prompts([
-			{
-				type: 'multiselect',
-				name: 'integrations',
-				message: 'Which frameworks would you like to use?',
-				choices: FRAMEWORKS,
-			},
-		]);
-		integrations = result.integrations;
-	}
-
 	spinner = ora({ color: 'green', text: 'Copying project files...' }).start();
 
 	// Copy
-	try {
-		emitter.on('info', (info) => {
-			logger.debug(info.message);
-		});
-		await emitter.clone(cwd);
-	} catch (err: any) {
-		// degit is compiled, so the stacktrace is pretty noisy. Only report the stacktrace when using verbose mode.
-		logger.debug(err);
-		console.error(red(err.message));
+	if (!args.dryrun) {
+		try {
+			emitter.on('info', (info) => {
+				logger.debug(info.message);
+			});
+			await emitter.clone(cwd);
+		} catch (err: any) {
+			// degit is compiled, so the stacktrace is pretty noisy. Only report the stacktrace when using verbose mode.
+			logger.debug(err);
+			console.error(red(err.message));
 
-		// Warning for issue #655
-		if (err.message === 'zlib: unexpected end of file') {
-			console.log(
-				yellow(
-					"This seems to be a cache related problem. Remove the folder '~/.degit/github/withastro' to fix this error."
-				)
-			);
-			console.log(
-				yellow(
-					'For more information check out this issue: https://github.com/withastro/astro/issues/655'
-				)
-			);
-		}
-
-		// Helpful message when encountering the "could not find commit hash for ..." error
-		if (err.code === 'MISSING_REF') {
-			console.log(
-				yellow(
-					"This seems to be an issue with degit. Please check if you have 'git' installed on your system, and install it if you don't have (https://git-scm.com)."
-				)
-			);
-			console.log(
-				yellow(
-					"If you do have 'git' installed, please run this command with the --verbose flag and file a new issue with the command output here: https://github.com/withastro/astro/issues"
-				)
-			);
-		}
-		spinner.fail();
-		process.exit(1);
-	}
-
-	// Post-process in parallel
-	await Promise.all([
-		...FILES_TO_REMOVE.map(async (file) => {
-			const fileLoc = path.resolve(path.join(cwd, file));
-			return fs.promises.rm(fileLoc);
-		}),
-		...POSTPROCESS_FILES.map(async (file) => {
-			const fileLoc = path.resolve(path.join(cwd, file));
-
-			switch (file) {
-				case 'CHANGELOG.md': {
-					if (fs.existsSync(fileLoc)) {
-						await fs.promises.unlink(fileLoc);
-					}
-					break;
-				}
-				case 'astro.config.mjs': {
-					if (selectedTemplate?.integrations !== true) {
-						break;
-					}
-					await fs.promises.writeFile(fileLoc, createConfig({ integrations }));
-					break;
-				}
-				case 'package.json': {
-					const packageJSON = JSON.parse(await fs.promises.readFile(fileLoc, 'utf8'));
-					delete packageJSON.snowpack; // delete snowpack config only needed in monorepo (can mess up projects)
-					// Fetch latest versions of selected integrations
-					const integrationEntries = (
-						await Promise.all(
-							integrations.map((integration) =>
-								fetch(`https://registry.npmjs.org/${integration.packageName}/latest`)
-									.then((res) => res.json())
-									.then((res: any) => {
-										let dependencies: [string, string][] = [[res['name'], `^${res['version']}`]];
-
-										if (res['peerDependencies']) {
-											for (const peer in res['peerDependencies']) {
-												dependencies.push([peer, res['peerDependencies'][peer]]);
-											}
-										}
-
-										return dependencies;
-									})
-							)
-						)
-					).flat(1);
-					// merge and sort dependencies
-					packageJSON.devDependencies = {
-						...(packageJSON.devDependencies ?? {}),
-						...Object.fromEntries(integrationEntries),
-					};
-					packageJSON.devDependencies = Object.fromEntries(
-						Object.entries(packageJSON.devDependencies).sort((a, b) => a[0].localeCompare(b[0]))
-					);
-					await fs.promises.writeFile(fileLoc, JSON.stringify(packageJSON, undefined, 2));
-					break;
-				}
-			}
-		}),
-	]);
-
-	// Inject framework components into starter template
-	if (selectedTemplate?.value === 'starter') {
-		let importStatements: string[] = [];
-		let components: string[] = [];
-		await Promise.all(
-			integrations.map(async (integration) => {
-				const component = COUNTER_COMPONENTS[integration.id as keyof typeof COUNTER_COMPONENTS];
-				const componentName = path.basename(component.filename, path.extname(component.filename));
-				const absFileLoc = path.resolve(cwd, component.filename);
-				importStatements.push(
-					`import ${componentName} from '${component.filename.replace(/^src/, '..')}';`
+			// Warning for issue #655
+			if (err.message === 'zlib: unexpected end of file') {
+				console.log(
+					yellow(
+						"This seems to be a cache related problem. Remove the folder '~/.degit/github/withastro' to fix this error."
+					)
 				);
-				components.push(`<${componentName} client:visible />`);
-				await fs.promises.writeFile(absFileLoc, component.content);
+				console.log(
+					yellow(
+						'For more information check out this issue: https://github.com/withastro/astro/issues/655'
+					)
+				);
+			}
+
+			// Helpful message when encountering the "could not find commit hash for ..." error
+			if (err.code === 'MISSING_REF') {
+				console.log(
+					yellow(
+						"This seems to be an issue with degit. Please check if you have 'git' installed on your system, and install it if you don't have (https://git-scm.com)."
+					)
+				);
+				console.log(
+					yellow(
+						"If you do have 'git' installed, please run this command with the --verbose flag and file a new issue with the command output here: https://github.com/withastro/astro/issues"
+					)
+				);
+			}
+			spinner.fail();
+			process.exit(1);
+		}
+
+		// Post-process in parallel
+		await Promise.all(
+			FILES_TO_REMOVE.map(async (file) => {
+				const fileLoc = path.resolve(path.join(cwd, file));
+				if (fs.existsSync(fileLoc)) {
+					return fs.promises.rm(fileLoc, {});
+				}
 			})
 		);
-
-		const pageFileLoc = path.resolve(path.join(cwd, 'src', 'pages', 'index.astro'));
-		const content = (await fs.promises.readFile(pageFileLoc)).toString();
-		const newContent = content
-			.replace(/^(\s*)\/\* ASTRO\:COMPONENT_IMPORTS \*\//gm, (_, indent) => {
-				return indent + importStatements.join('\n');
-			})
-			.replace(/^(\s*)<!-- ASTRO:COMPONENT_MARKUP -->/gm, (_, indent) => {
-				return components.map((ln) => indent + ln).join('\n');
-			});
-		await fs.promises.writeFile(pageFileLoc, newContent);
 	}
 
 	spinner.succeed();
 	console.log(bold(green('✔') + ' Done!'));
 
-	console.log('\nNext steps:');
-	let i = 1;
-	const relative = path.relative(process.cwd(), cwd);
-	if (relative !== '') {
-		console.log(`  ${i++}: ${bold(cyan(`cd ${relative}`))}`);
+	const installResponse = await prompts({
+		type: 'confirm',
+		name: 'install',
+		message: `Would you like us to run "${pkgManager} install?"`,
+		initial: true,
+	});
+
+	if (!installResponse) {
+		process.exit(0);
 	}
 
-	console.log(`  ${i++}: ${bold(cyan('npm install'))} (or pnpm install, yarn, etc)`);
-	console.log(
-		`  ${i++}: ${bold(
-			cyan('git init && git add -A && git commit -m "Initial commit"')
-		)} (optional step)`
-	);
-	console.log(`  ${i++}: ${bold(cyan('npm run dev'))} (or pnpm, yarn, etc)`);
+	if (installResponse.install && !args.dryrun) {
+		const installExec = execa(pkgManager, ['install'], { cwd });
+		const installingPackagesMsg = `Installing packages${emojiWithFallback(' 📦', '...')}`;
+		spinner = ora({ color: 'green', text: installingPackagesMsg }).start();
+		await new Promise<void>((resolve, reject) => {
+			installExec.stdout?.on('data', function (data) {
+				spinner.text = `${installingPackagesMsg}\n${bold(`[${pkgManager}]`)} ${data}`;
+			});
+			installExec.on('error', (error) => reject(error));
+			installExec.on('close', () => resolve());
+		});
+		spinner.succeed();
+	}
+
+	const astroAddCommand = installResponse.install
+		? 'astro add --yes'
+		: `${pkgManagerExecCommand(pkgManager)} astro@latest add --yes`;
+
+	const astroAddResponse = await prompts({
+		type: 'confirm',
+		name: 'astroAdd',
+		message: `Run "${astroAddCommand}?" This lets you optionally add component frameworks (ex. React), CSS frameworks (ex. Tailwind), and more.`,
+		initial: true,
+	});
+
+	if (!astroAddResponse) {
+		process.exit(0);
+	}
+
+	if (!astroAddResponse.astroAdd) {
+		ora().info(
+			`No problem. You can always run "${pkgManagerExecCommand(pkgManager)} astro add" later!`
+		);
+	}
+
+	if (astroAddResponse.astroAdd && !args.dryrun) {
+		await execaCommand(
+			astroAddCommand,
+			astroAddCommand === 'astro add --yes'
+				? { cwd, stdio: 'inherit', localDir: cwd, preferLocal: true }
+				: { cwd, stdio: 'inherit' }
+		);
+	}
+
+	const gitResponse = await prompts({
+		type: 'confirm',
+		name: 'git',
+		message: 'Initialize a git repository?',
+		initial: true,
+	});
+
+	if (!gitResponse) {
+		process.exit(0);
+	}
+
+	if (gitResponse.git && !args.dryrun) {
+		await execaCommand('git init', { cwd });
+	}
+
+	console.log(`\n${bgCyan(black(' Next steps '))}\n`);
+
+	const relative = path.relative(process.cwd(), cwd);
+	const startCommand = [];
+	if (relative !== '') {
+		startCommand.push(bold(cyan(`cd ${relative}`)));
+	}
+	if (!installResponse.install) {
+		startCommand.push(bold(cyan(`${pkgManager} install`)));
+	}
+	startCommand.push(bold(cyan(pkgManager === 'npm' ? 'npm run dev' : `${pkgManager} dev`)));
+	console.log(startCommand.join(' && '));
 
 	console.log(`\nTo close the dev server, hit ${bold(cyan('Ctrl-C'))}`);
-	console.log(`\nStuck? Visit us at ${cyan('https://astro.build/chat')}\n`);
+	console.log(`Stuck? Visit us at ${cyan('https://astro.build/chat')}\n`);
+}
+
+function emojiWithFallback(char: string, fallback: string) {
+	return process.platform !== 'win32' ? char : fallback;
+}
+
+function pkgManagerFromUserAgent(userAgent?: string) {
+	if (!userAgent) return 'npm';
+	const pkgSpec = userAgent.split(' ')[0];
+	const pkgSpecArr = pkgSpec.split('/');
+	return pkgSpecArr[0];
+}
+
+function pkgManagerExecCommand(pkgManager: string) {
+	if (pkgManager === 'pnpm') {
+		return 'pnpx';
+	} else {
+		// note: yarn does not have an "npx" equivalent
+		return 'npx';
+	}
 }