+++ packages/markdown/remark/src/index.ts
@@ -1,4 +1,4 @@
-import type { MarkdownRenderingOptions } from './types';
+import type { MarkdownRenderingOptions, MarkdownRenderingResult } from './types';
 
 import createCollectHeaders from './rehype-collect-headers.js';
 import scopedStyles from './remark-scoped-styles.js';
@@ -22,11 +22,13 @@ import rehypeRaw from 'rehype-raw';
 export * from './types.js';
 
 export const DEFAULT_REMARK_PLUGINS = ['remark-gfm', 'remark-smartypants'];
-
-export const DEFAULT_REHYPE_PLUGINS = ['rehype-slug'];
+export const DEFAULT_REHYPE_PLUGINS = [];
 
 /** Shared utility for rendering markdown */
-export async function renderMarkdown(content: string, opts: MarkdownRenderingOptions) {
+export async function renderMarkdown(
+	content: string,
+	opts: MarkdownRenderingOptions
+): Promise<MarkdownRenderingResult> {
 	let { mode, syntaxHighlight, shikiConfig, remarkPlugins, rehypePlugins } = opts;
 	const scopedClassName = opts.$?.scopedClassName;
 	const isMDX = mode === 'mdx';